[[ Bug 19474 ]] Make sure unnamed objects use id form
This patch ensures that when a script object requests the long
name of an object, it uses the id form when an object is unnamed.

The problem was caused by adding a new name mode (LONG_NAME_NO_FILENAME)
but not testing for that when deciding what form of name to use for
an unnamed object.