[[ Toolchain ]] Generate shims with uniform signature for binding array
This patch changes the type of the ordinal binding array from void*
to void (*)(void *rv, void **av). The latter type means that the
function can be called directly with the argumenst as prepped for
ffi_call, but without using ffi_call.

Shims are generated which call the actual function, taking arguments
from the av array, and putting the result in rv.

In order to support generating the shims, query methods have been
added to the ScriptBuilder API which compute the 'primitive' (C) types
of things in the handler signature. This allows C code to be generated
which calls the function correctly.

This approach is needed as it is much better not to use function pointer
conversions in Emscripten and this approach eliminates the need. Additionally
it gives us an extra point of checking of (C) foreign handlers as at link
time, emscripten will emit errors if the signature of a function in one
place is inconsistent with another.