[[ VM ]] Use builtins array if present for ordinal bindings
This patch adds the idea of a builtins array to a ScriptModule
and changes how binding strings are resolved for foreign
handlers.

If the module containing the handler definition has a builtins
array and the binding string is an integer, then it will use
the pointer in the builtins array as the function pointer when
resolving.