Update if.lcdoc
Further modifications to summary ensuring if is properly marked up with <if> to ensure it will be shown in bold in dictionary,  add explanation of requirement for 'end if'  keyword in the summary.  Added reference to 'else if' and a link to that.  There is no "elseIf.lcdoc" file for those to reference however.  I have submitted a bug report #19786 - Dictionary does not contain 'else if' keyword entry.