[[ Bug 19783 ]] Don't retain errors from other events in try
This patch ensures that if statements are executed successfully that
there are no errors on the error stack. This may happen if the error
was from a different event and not thrown due to the trylock.

(cherry picked from commit ba6708563b69b5d13b85bc50ea9f61dab9e5bc59)