[[ Emscripten Networking ]] Added support for load command to HTML5.
Support for the the load command, similar to that in the mobile
engines has been added for HTML5.

The code was pretty much all there, it just needed very minor tweaking:
A param was missing in the progress callback and the timeout was being
set incorrectly, causing most URLs to fail.