[[ Bug 19837 ]] Use MCObjectHandle for objectrefs in pending messages
This patch uses MCObjectHandles to manage the objects in the
pending message queue.

As the pending message queue on this branch does not conform to
C++ array requirements, and other reasons; rather than use
MCObjectHandle as a direct member of MCMessageList, it uses
MCObjectHandle*, and creates a heap-based instance. This means
that the lifetime management of it matches that of the name
field exactly.