[[ AndroidService ]] Add a background service to Android and example
This patch adds the infrastructure to allow LCB modules to start
and stop a background service on Android. This allows operations to
be performed whilst the engine is in the background.

A long with the basic infrastructure, a simple example which uses
a MediaPlayer to play audio in the background is also included.