[[ WidgetContext ]] Update widget context on entry/exit to widget handlers
This patch moves the code which stacks up the current widget state
from widget-ref.cpp to callbacks invoked by libscript.

Before a widget handler is called, the Entry callback is called
allowing MCcurrentwidget and the execution lock to be taken.

After a widget handler is called, the Leave callback is called
allowing MCcurrentwidget to be reset and execution lock to be
removed.