[[ AndroidButton ]] Temporary fix to thread affinity and android button
This patch implemented a temporary way of expressing thread affinity
when binding to java method on android.

If a method of an Android Java object requires that it be run on
the UI thread, then the binding string should use 'javaui' and
not 'java'.

Any handlers bound with 'javaui' will ensure that the engine is
running on the Android UI thread before calling them, and return
to the original thread thereafter.

Furthermore, all methods invoked using interface proxy callbacks
(via the LCBInvocation proxy) will ensure they are called on the
engine's script thread.

The button example has been modified to use 'javaui' in appropriate
places, and the label property has been renamed to 'label'.