[[ Bug 19873 ]] Ensure MCForeignValue content is 64-bit aligned
This patch forces the content portion of MCForeignValue structs
to be 64-bit aligned. This ensures that emscripten does not have
to generate 'special' code to access a double which has been
packed into the struct as when it does use the non-aligned
code path, the numbers gain error which cause repeat loops to
behave unexpectedly.