[[ Bug 19936 ]] Always create a window mask on Mac
This patch fixes a bug which prevented using an image with no
transparency as a window shape. This is because it was not
creating a mask on mac if the image had no transparency, an
artifact from the port from Cocoa where (prior to that) an
image with no transparency would use a region based window mask
which Cocoa doesn't support.