[[ Tests ]] Add TestSkipIf[Not] command
This patch adds `TestSkipIf` and `TestSkipIfNot` commands to
simplify the handling of particular features that are unupported
in a given environment. This allows us to change the condition
for skip centrally. For example, should Emscripten start supporting
wait we could turn all those tests on with a one line change.