[[ FFI ]] Tidy up libscript FFI support
This patch tidies up the current implementation of FFI in libscript.

The type of foreign handler is now determined by the 'language'
field. This is set after successful parsing of the binding string and
can be one of 'C', 'BuiltinC' or 'Java'.

Parsing of the binding string has been split up into per-language
functions. The language prefix of the string is first parsed out, and
that is used to determine which further function to call.

The C++ and ObjC related errors and bindings have been removed as they
are both to be implemented via C bindings and therefore the distinction
is unnecessary.