[[ LCB ]] Add varargs support to foreign handlers
This patch adds varargs support to foreign C handlers, allowing
binding to a wider range of C functions.

The syntax for vararg handlers is similar to that of C:
```
  foreign handler myvarargs(in pFirst as Pointer, ...) returns CInt binds to ...
```
Using the token `...` to indicate the start of the variable
parameter list.

To implement vararg functions, a new parameter mode 'variadic'
has been added which causes a handler to be treated as variadic if
such a parameter with a mode is present.

As C requires any int type with rank less than int to be promoted
to int, and float to be promoted to double (when passed to a non-fixed
parameter in a variadic function), a 'promotedtype' and 'promote'
function have been added to foreign type descriptors to describe
the appropriate relation.