[[Bug 18305]] Add Synonyms to Dictionary (#5669)
LC 7 allowed searching for synonyms within the dictionary.  Implementation appears to have been by duplicating the dictionary entries for each synonym.  This fix will duplicate the dictionary entry for each synonym when the entries are built from the repository.  "Synonym of" is added to the display name of each non-glossary entry (glossary synonyms do not seem to mean that the terms mean the same thing - i.e. card button, card field, card control, card object).

The change is made in the "revDocsParseDictionaryToLibraryArray" handler.  On each pass through the loop, the resulting array is tested to see if the "Synonyms" key contains data.  If it does, then loop through each synonym and duplicate the entry (taking into account that glossary entries specify themselves as a synonym).  The display name is changed to the synonym and "Synonym of ..." is added as a cross-reference for non-glossary entries.