[[Bug 19622]] Handle line continuations
Line continuation characters are not handled in the main loop that adds lines to "tData".  They do get parsed for "syntax" type (gets all of "tData" - revDocsUpdateModularSyntaxDocBlocks), but "handler", "function", and "command" types only get passed the first line of "tData" (revDocsParseHandler, revDocsUpdateDocBlocks).  Proposed solution is that while in an entry, check the previous line for the continuation character.  If present, remove it and the return.