[[ Bug 19894 ]] Render caret with the correct blend mode
The caret has always been rendered using bitwise-xor blend
mode when rendering against an opaque background. However,
since the move to the new version of Skia we have had to
remove the 'legacy' blend modes.

Fortunately, when rendering white or black, using exclusion
or difference blend modes gives the same result as bitwise-
xor.

As exclusion is actually closer to bitwise-xor in general
(mathematically at least), this patch changes the use of
GXxor to GXblendExclusion in MCField::drawcursor.