[[ Bug 16131, 17323 ]] Linux backdrop issues
This patch fixes two linux backdrop issues:

- Bug 16131 - `gdk_window_fullscreen` is not guaranted to fullscreen
the window and does nothing if the window is already fullscreen. On
Ubuntu this was showing the behavior of fullscreening the first time
and showing as the window size on subsequent times
- Bug 17323 - `gdk_window_show_unraised` is being used to show the
backdrop but there is no guarantee the backdrop window is already
above other windows. This patch lowers the window before it is shown
unraised to ensure it is the lowest window in the application.