[[ Bug 20055 ]] Support building commercial extensions
This patch adds an option to the extension builder to remove the source
file after building so that commercial extensions do not include the
source.

This patch also comments out the output directory for extensions which
was casuing the commercial target to not run as it thought the output
was satisfied. This change means extensions are rebuilt every time
but as it reduces the issues we have with branch changes that is probably
a reasonable trade.