[[ Bug 17639 ]] Fix vertical placement of caret on wrapped lines
This patch fixes the vertical 'drift' which previously occured
vertical when positioning the caret on soft-wrapped lines of a
paragraph.

The issue was caused by 'getsplitcursorrect' using floor, compared
to 'getcursorrect' using ceil - the two methods used essentially
the same code beyond that (hence why the error occurred in the first
place).

The getsplitcursorrect method has been removed, and the getcursorrect
method now has an extra parameter which specifies either full, primary
or secondary cursor. The cursor type parameter has a default of 'full'.