[[ Bug 17098 ]] Ensure there is always a segment for each tab
This patch ensures that if the last block on a line has a trailing
tab, then an empty block is added at the end so that an empty
segment is created.

This fixes the placement of the caret for non-left aligned tabs
when the last tab is empty.

At some point in the past an empty block was always created for
an empty tab segment - however this caused bug 13887. Only creating
the empty block for the very last segment seems to not reintroduce
that bug.