[[ Perf ]] Special case strict long ids
This patch is an experiment to work out what the overhead of parsing
long ids is.

It is very much WIP, and not complete - in particular, it does not check
ownership of controls in the control chain; and is hardcoded to only work
for this card of the target stack.

However, it does seem to suggest that long ids could be parsed and evaluated
in the same time it takes for a pre-compiled chunk expression to be evaluated
which is promising (the speed up is nigh on a factor of 10, on my machine).