Use MCInterfaceTryToEvalStackFromString directly
The code that opens a stack from the file system uses MCInterfaceTryToEvalStackFromString.  Since the input is a MCStringRef, it would be a drop-in replacement.

If that call failed to retrieve a stack, then MCInterfaceTryToEvalStackFromString should fail too.  I'm not sure if it would be proper to directly set the error message after the MCInterfaceStringCouldBeStack test.  Worst case, the file is evaluated a second time before the error is returned with the nil pointer.