[[ Bug 17810 ]] Non-LF line endings in script only stacks
No code changes were necessary for LC9 to support reading files with the various line endings.  Separate PR outstanding for BOM handling.
Added 2 bool flags to the stack struct to retain state of line endings used when importing a script only stack.  LF is set to true upon object creation as the default for new script only stacks.  (Could use an #if defined to set CRLF for Windows if desired)
Added methods to get/set the line ending values on a stack object.
[Considered using MCSOutputLineEndings enum (osspec.h), but Monte mentioned it would need to get moved to foundation first.  Changing the code would not be very difficult though.  kMCSOutputLineEndingsNative is not really correct for Mac since __CR__ is still used and I understand can't be changed right now.  If we changed to this enum, then the default would need to be LF except for Windows.]
When the stack object is created for a script only stack, the raw script is checked for CR and LF line endings.  If neither found, then LF is selected.
MCStringConvertLineEndingsFromLiveCode was overloaded with a version that accepts the 2 bool values for CR and LF endings.  The save stack call was modified to use this version with the line ending values from the stack object.