[[ Notes ]] Refactor release note collation
This patch moves all release notes to the V2 format and implements
some support for V1 style notes in the V2 note collation to ease
the transition.

Additionally, this patch reads release notes from component folders
in the public and private repos. Component folders contain a YAML
metadata file and may include a notes, dictionary and glossary
directory in addition to code and other resources related to the
component. It is intended that the current content of the extensions
folders be ported to components as time allows.