[[ Bug 20286 ]] Add tool property of image - EXPERIMENT
This patch is an experiment to see how well a tool property of images
would work.

As posed, it is quite straight-forward, but there are some details to
pin down. For example, when should the image's tool be used? How should
the IDE use it in edit mode - should it?

At the moment tool property of an image can only be 'pencil' and that
overrides any global tool other than 'help'. This means that if an
image has its tool property set in the IDE, in both browse and edit
mode the image is editable.

Also, cursor changes don't currently occur which makes it even more
confusing. (Although this is just an experiment - so I've not poked
too deeply!)