[[ Bug 20308 ]] Fix 64-bit Mach-O section writing
This patch ensures that section_64 is used when updating the project
section in a 64-bit Mach-O deployment. Previously, section was being
used which resulted in the fields being slightly wrong. The latter has
never affected execution, but has started to affect tools such as
'otool' which are now doing more validation.