[[ Bug 18955 ]] Fix crash when using file dialogs from browser widget
This patch fixes a crash when using the HTML file dialog from within
a browser widget or revbrowser.

The crash was caused by releasing an auto-released object, rather than
ensuring it the object was created in an auto-release pool.

The direct category addition to NSObject in revbrowser for the WebUIDelegate
methods have been removed and placed in the adapter class in revbrowser
instead and the necessary code has been replicated in MCWebUIDelegate for the
Mac browser implementation in libbrowser ensuring the two can work
separately.