Restore "new folder" button on Mac folder dialog
When the file/folder routine was abstracted (commit 93b3d28), the "new folder" button was removed from the folder dialog.  This commit restores the code line and original comment where it would have landed in the updated function.  I originally branched this off of develop due to the missing save dialog, but due to this regression it should probably be applied to 8.1.7 as well.