[[ Bug 20325 ]] Add notion of foreign type constructor
This patch adds the notion of a 'foreign type constructor'. This
expands the foreign type clause in lcb to allow a binding string of
the form "<name>:<args>". Here, name is resolved as a C function with
prototype bool (*)(MCStringRef, MCTypeInfoRef&). Where the first
argument is passed the args part of the binding string. The function
should create a new typeinfo based on the arguments it is given.