[[ Bug 20363 ]] Always add DIB format to clipboard for images
This patch ensures that the CF_DIBV5 format is always added when
a PNG, GIF or JPEG image is added to the clipboard on Windows.

Note: This breaks the idea of the raw clipboard slightly as it
means formats are not entirely separable at that level. However
one can work around that by setting the DIB format *after* the
PNG, JPEG or GIF format. Ideally, synthesized formats would be
added at clipboard unlock and derived from the explicitly set
formats and only if they are not explicitly set.