[[ Bug 19909 ]] Add emscripten deploy library.
An internal command has been added to the Mac engine to return a
list of all the installed web browsers. We get a similar list on
Windows by querying the registry. For Linux we check a static
list of browsers to see what's installed.

The list of browsers is used as deploy targets for running HTML 5
standalones.

The deploy library creates an instance of a local HTTP server using
the HTTPd library, builds the standalone then launches the chosen
browser, pointing it to the appropriate localhost path.