Revert "[[Bug 18305]] Add Synonyms to Dictionary (#5669)"
This reverts commit d72f1632e7302c856246143942437b981c88c854.
Due to a regression where search terms such as "mobile" will display the "iphone" synonyms first and give the impression that those are the preferred entries.  Changes are being made to the IDE (dictionary viewer) to implement the synonym search there.

This code also caused issues with duplicates being displayed in the new Autocomplete feature.