[[ Bug 20504 ]] Fix __MCStringCheck function
This function corrects the __MCStringCheck function when the
input string contains a decomposed char sequence which can
be mapped to a composed char in the native encoding.

Additionally, it corrects the ISO8850-1 'native pair' mapping
table ensuring that decomposable native chars in the encoding
map back to their composed counterparts.