[[ Bug 20504 ]] Don't assume Native => Trivial
This patch further changes the __MCStringCheck function and also
adjusts the __MCStringSetFlags function.

Previously, there was an assumption that Native => Trivial, which
is not true in the case of there being a native composable pair
(e.g. e,acute). The __MCStringCheck function now scans the entire
string and collects whether the string is Basic, Trivial, and/or
Native. The __MCStringSetFlags function no longer assumes that
Native => Trivial.