[[ GoogleTests ]] Turn off C++11 support
This patch ensures that googletest doesn't compile with C++11
features which (at least when compiling develop-8.1 on newer
versions of Xcode) seem to be enabled even though the library
support isn't there.

This change can be reverted after merging up to develop - where
we should be able to enable such support.