[[Bug 13825/18097]] Profile Manager/Geometry Manager on Mobile
Profile Manager (PM) and Geometry Manager (GM) do not work on mobile due to missing library code in the standalone that is built.

GM depends on a handler present in the common library (REVTargetStack).  That library was not being loaded on mobile.  There is a good bit of code in that library that is not applicable to mobile.  There is a separate PR to add a mobile library with just the applicable code.

PM depends on getprop/setprop handlers in the common library.  Same solution as for GM.

The code that included the PM library itself was not included in the standalone builder scripts for mobile.  That code is included in this PR for both iOS and Android.