[[ Bug 20462 ]] Fix crash when undo group deletion
This patch fixes a problem when a group is deleted, and then undo
is performed on it.

Deleting an object now causes it's object handle to be cleared
and removed. This means that all descendent object's parent
references are cleared.

When undoing, the control which is temporarily owned by the undo
list must be reinsterted into the object tree which requires
setting the parent references correctly.

This is done in MCControl::undo by using a visitor which visits
a control to recreate its weak proxy, and then visits its children
to set the parent link.