[[Bug 20703]] revDocsParseDescriptionOfEnum continuation line parse error
When enum description lines continued, the additional lines wound up getting added to the description.  The 2 sort docs showed the issue for the sortType of numeric.  The variable was already defined in the function to perform the check, but was not implemented inside the repeat loop.

This update will allow docs with enums to be wrapped according to the specs (some that I checked that looked good in the release were all on one line even when much longer than 72 characters).