[[ Bug 20584 ]] Utilise SB library helpers when building html5 Much of the deploy param computation was unnecessarily duplicated in the script of stack revSaveAsEmscriptenStandalone. This patch utilises as many of the SB library helper functions as possible.
At the moment, we need to retain the custom code for extensions
because on all other platforms they are compiled into the
executable. Also library mappings (which are used on other platforms
to map extensions' resource paths) are not supported. They are not
currently needed, but some form of support will  be needed if and
when side module support is added to the emscripten engine.