[[ Bug 19358 ]] Split font name string correctly
This patch corrects a flaw in the mobile font creation code which
would attempt to split the name at a comma without checking there
was actually a comma.

In release mode builds, this caused the font name to be empty in
many cases (due to clamping in libfoundation), meaning fonts
would not appear as set.