[[ Bug 19543 ]] Dictionary improperly parses hard wrapped lines where word is followed by colon
When a paragraph ends with a colon and the last word falls on a new line then that word is parsed as a new element.  Due to the way the regex is parsed, something inside parenthesis before the colon could also trigger this bug.

The proposed fix:  Unless there is a preceding blank line, assume that this term is part of the existing content.  If there is no content, then assume it is indeed a new element.  An exception is if the element contains the content on the same line, in that case a new element can start on the next line.  Some LCB documentation has "Description:" immediately following another multi-line element without a blank line (i.e. examples).  Added a check that can be extended if other elements are found that need to be treated as an exception.

I ran a diff on the built_api.js from before my change and after.  There are 584 lines of changes.  Not all of them are evident in the dictionary viewer.  (Only the first 128 lines are for ~32 actual changes, the rest were due to reordering of svgpath and svgspinner in the JSON.)

Some entries fixed by the change:
- open printing
- open printing to pdf
- switch