[[ Graphics ]] Paint Objects
This patch changes the distinct paint attributes in MCGContextRef
to be a new MCGPaintRef type.

The MCGPaintRef type inherits from an ref-counted MCGObjectRef type
which is implemented as a partially concrete C++ class.

There are specific paint types for solid color, patterns and
gradients.

Filling or stroking can be disabled by setting the relevant paint
to none.