[[ Graphics ]] Paint Opacity
This patch implements the currently unimplemented fill and
stroke opacity properties.

The fill and stroke opacity differ from opacity in that they
are applied as attenuations on the pixels generated by the
paint, rather than acting as a mask for composition to the
target surface.

In Skia, paints which use shaders take the opacity from the
alpha of the color which is set on the SkPaint. For colors,
the alpha value must be multipled by the opacity.