[[ CodeResources ]] Impmement `revLibraryMapping` property
This patch implements a `revLibraryMapping` array property to get and
set the mapping path for a library. This allows the IDE to have full
control over the where to find the libraries which is required because
on some systems libraries must be placed in certain locations rather
than inside the extension package.