[[ CodeResources ]] Allow full paths for library mappings
This patch allows full paths to be used for library mappings by checking
if the mapping is prefixed with `./` we can determine if it is intended
to be a path relative to `MCappcodepath`. If it is not prefixed then
the mapping path is passed directly to `MCS_resolvepath`. Currently
all relative mappings are already prefixed by `./` so this is a
backwards compatible change.