[[ CodeResources ]] Implement example extension with code resource
This patch implements a library extension that includes a code resource.
The extension is an INI to LiveCode array parser using the inih library.

As extensions are not built when building mobile platforms this patch
adds a target for cross platform extension dependencies that will be
built for all platforms.