[[ Bug 11039 ]] Ensure set executing behavior to empty throws
This patch ensures that setting the behavior of an object to empty while
the behavior script is executing throws an error. A previous patch ensured
that setting to a sifferent object throws. If no error is thrown and there
is an error thrown in the message path then the invalid behavior is still
in the execution contexts causing a crash.