Line Ending Conversion Corrections
In two cases, native Mac line endings are not being converted (CRLF and LFCR are converted to LF but CR was not touched).
In one case, the constant CR was used where it should have been numToCodepoint(13).
Changed to use the LF constant to make it more readable.
Also changed numToChar to numToCodepoint since the former function is deprecated.