[[ Bug 20578 ]] Fix replaceText performance regression
In order to fix Bug 19005 it was necessary to perform a UnicodeCopy
on the input string to MCR_exec which is used by the replaceText()
function.

However, the replaceText function calls MCR_exec multiple times
which means that complete copy of the input string is being
performed many times unnecessarily.

This patch performs a UnicodeCopy of the string before the call
to MCR_exec, meaning it is only copied once for an entire replaceText
operation.