[[ Graphics ]] Drawing
This patch adds the notion of a 'Drawing' to libgraphics.

A drawing is a simple bytecode format for efficiently representing
what are generally called 'vector graphics' - i.e. a combination
of geometric shapes which are painted using solid colors, gradients
and patterns.

Currently the format can represent a subset of SVG Tiny 1.2, although
it is designed so that it can be easily evolved to support ever
greater subsets.

The drawing format has been designed so that a drawing file can be
mapped into memory and executed without any processing - greater
efficiency being possible in the future by caching various runtime
aspects.