[[ ObjcDelegates ]] Allow creation of delegates based on informal protocols When an objective-c protocol's methods are all optional, its information is not available using the objc runtime API - obj_getProtocol returns nil.
In order to allow creation of delegates which implement methods from these
protocols, this patch adds MCObjcCreateInformalDelegate and
MCObjcCreateInformalDelegateWithContext functions. These take a list of
foreign handler objects which are used to create a mapping from selector
names, allowing the construction of the NSInvocation method signature
required to forward the invocation to the target object.

Since there is a lot of shared code, the previous ObjcDelegate class is
used as a base class for the ObjcFormalDelegate class (where a Protocol
object is available) and ObjcInformalDelegate (where it is not).