[[ Stubs ]] Use MCLog in generated stubs files
This patch changes the stubs file maker to use `MCLog` instead of
`fprintf` to `stderr`. This is mainly helpful on Windows where
`stderr` is not sent to the debugger output console. However, on
all platforms it has the advantager of including the file name
and line number in the output automatically.