[[ CodeResources ]] Remove options section from gyp platform_id var
This path removes the options section from the platform_id variable on
windows. There is an issue that the configurations are resolved last and
therefore the CRT mode can not be determined before the variable needs
to be used. A TODO note as been included to add the options section when
possible.