[[ LCBLicenseCheck ]] Add license checking syntax
This patch adds new license checking syntax to LCB in a
com.livecode.commercial.license module.

This module's implementation lies in commercial, but the syntax
definitions are currently in the community repo as there is no
commercial-specific lc-compile as yet.

As syntax implementation methods are only bound if used,
the commercial module is only required if the license syntax is
used - which means it does not need to be included in community.

This patch adds license checking and ensuring syntax to
LCB allowing modules to ensure they are being run in an
engine they are licensed to support. Implementation is
in com.livecode.commercial.engine.