[[ SVG ]] Improve paint server support
This patch improves paint support in the drawing format and svg compiler.

It adds the full range of named colors from SVG1.1 (now specified in the
svg specification file).

It adds support for percentage values in gradient coordinates and radii
for both the object bbox case and the viewport case, where the current
viewport is fixed size.

Finally it adds support for conical gradients - i.e. radial gradients
which have a focal point and optional focal radius.