[[ SVG ]] Further Integration
This patch improves the support for drawing-based images.

It ensures that drawing-based images render correctly when
an image has a transformation (rotation now works).

It ensures that drawing-based images render correctly when
used as icons (in particular, with iconGravity).

It ensures that drawing-based images work when used from
LCB via the 'image from data' constructor.