[[ SVG ]] Static Viewport
This patch changes the drawing format and svg compiler so that a
fixed width/height is *always* present in the header - this much
better reflects that drawings are currently entirely static and
simplifies the compiling of SVG files in that form. Additionally
it also means that a suitable size is always available - which is
important for drawings to behave like images, as images require
a 'pixel size' for various operations in the engine.