[[ SVG ]] Simplify Opcodes
This patch makes a number of changes to the drawing format
opcodes to simplify it.

The relative path commands have been removed as the relative
commands don't offer any advantage after compilation than
their non-relative counterparts. This change reduces the
number of path opcodes by half, allowing future use of the
top 4-bits of the opcode for a repeat count.

The paint and transform opcodes now define a single object,
rather than a sequence. This better reflects the current
static nature of the drawing format, where transforms can
always be flattened to a single matrix and there can only
be one paint defined at a time.

Finally, a RoundedRectangle opcode has been added, better
distinguishing between a rectangle with or without rounded
corners.