iconSVG library - current family
Add the ability to get/set the current icon family in the iconSVG library.  The initial value is "fontawesome" (previous behavior).  The "ide" family is loaded when the library is initialized and is available immediately.  If the user adds additional families, they will also be available.  This will allow the existing `iconpicker` to be used to view icons from any family that is loaded.