[[ TileCache ]] Selection Layer
This patch ensures that all selection related decorations (i.e.
size handles and edit-tool handlers) are drawn in the card's
foreground layer, and any changes to their appearance causes
that layer to be dirtied in the appropriate places.

This fixes a number of rendering issues with selection handles
both in accelerated rendering mode and outside of it.