[[ Bug 20856 ]] Enable sub-pixel positioning in CoreText rendering
This patch ensures that the 'setSubpixelText' flag is set to true
in the Skia paint used to render text on the CoreText implementation
of libgraphics text operations.

This change only effects the visual presentation of text - the metrics
(as seen by anything higher up) is unchanged. Previously Skia would
have been aligning the individual glyphs to device pixel boundaries
resulting in a subtle misplacement, noticeable when comparing with
the same text rendered in other CoreText using applications.