[[Bug 19206]] Clipboard always converts plain text to styled text
Whenever `text` was placed on the clipboard, it was first converted to LiveCode styled text and then put on the clipboard as styled text, RTF, HTML, and plain text.  This introduced errors when pasting to other applications since they would prefer the HTML version which made the text appear double spaced.

Updated `MCClipboard::AddText` such that is will clear the clipboard and only place the plain text representations on the clipboard.  The clipboard needed to be cleared to avoid the HTML/RTF from a previous copy being left on the clipboard which would cause issues when attempting to paste.

Updated related documentation to clarify that placing styled text on the clipboard will convert to the other representations, but plain text will not.  Using the copy text menu command will take styled text as will the copy command (chunk of field syntax).