[[ Bug 20933 ]] Fix unchecked codegen overflow for large handlers
This patch changes the way instructions are turned into bytecode
for a handler. Previously there was an unreported hard limit which
meant that any handler which contained jumps with a delta > 8192
would cause a bad module to be generated. This has been fixed
by trying different sizes for jump delta encodings until one is
found which works. Very small handlers will be encoded in fewer
bytes than they are now, handlers which currently compile and work
will be the same size, and large handlers will now compile and
work correctly.